+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+development:
+  * add SQL logger to development environment
+
 0.3.0:
 
   * remove Melodiest::Auth::Http module, it's better to use sinatra-authorization extension